backend/requirements.txt
anyio==4.11.0
asgiref==3.10.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
coverage==7.11.1
cryptography==46.0.3
curl_cffi==0.13.0
Django==5.2.8
django-cors==0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==37.12.0
firebase==4.0.1
firebase_admin==7.1.0
frozendict==2.4.6
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jwcrypto==1.5.6
jws==0.1.3
lxml==6.0.2
msgpack==1.1.2
multitasking==0.0.12
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
pip-review==1.3.0
platformdirs==4.5.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.11.0
wikipedia==1.4.0
yfinance==0.2.66
SPARQLWrapper==2.0.0
PyGithub==2.4.0
setuptools>=65.0.0
language-tool-python==2.8
Jinja2==3.1.4
MarkupSafe==3.0.2
python-docx==1.2.0
reportlab==4.4.4
celery==5.3.5
django-celery-beat==2.8.1
redis==4.5.4
